ledger(s1): guard 0040 on absent stream tables, admit fractional decision-fact instants, single retired-table owner (#930)

+++ packages/protocol/src/ledger/schema.ts
@@ -8,7 +8,7 @@ export namespace DecisionFact {
       key: z.string().min(1),
       type: z.string().min(1),
       data: PlainObjectSchema,
-      timeCreated: EpochMs.int().nonnegative(),
+      timeCreated: EpochMs.nonnegative(),
     })
     .strict();
   export type Record = z.infer<typeof Record>;
